Better structure.
